Day 9 Solution. IntCodeVM now uses Longs and supports relative mode. Had to adjust previous solutions due to int -> long.

+++ Solutions/Year2019/Day02/Solution.cs
@@ -18,7 +18,7 @@ namespace AdventOfCode.Solutions.Year2019 {
             vm.RunProgram();
 
             /* read out memory 0 */
-            int result = vm.ReadMemory(0);
+            long result = vm.ReadMemory(0);
 
             return result.ToString(); 
         }
